feat: agency mgmt (#6277)

+++ api/src/services/agency.service.ts
@@ -20,6 +20,7 @@ import { SuccessDTO } from '../dtos/shared/success.dto';
 import { IdDTO } from '../dtos/shared/id.dto';
 import { buildOrderBy } from '../utilities/build-order-by';
 import { OrderByEnum } from '../enums/shared/order-by-enum';
+import { buildFilter } from '../utilities/build-filter';
 
 @Injectable()
 export class AgencyService {
@@ -63,23 +64,40 @@ export class AgencyService {
   buildWhereClause(params: AgencyQueryParams): Prisma.AgencyWhereInput {
     const filters: Prisma.AgencyWhereInput[] = [];
 
-    const jurisdictionId = params?.filter?.find(
-      (filter) => filter[AgencyFilterKeys.jurisdiction],
-    )?.[AgencyFilterKeys.jurisdiction];
-
-    if (jurisdictionId) {
+    if (params.search) {
       filters.push({
-        jurisdictions: {
-          is: {
-            id: jurisdictionId,
-          },
+        name: {
+          contains: params.search,
+          mode: Prisma.QueryMode.insensitive,
         },
       });
     }
 
-    return {
-      AND: filters,
-    };
+    if (!params?.filter?.length) {
+      return { AND: filters };
+    }
+
+    params.filter.forEach((filter) => {
+      if (!filter[AgencyFilterKeys.jurisdiction]) return;
+
+      const builtFilter = buildFilter({
+        $comparison: filter.$comparison,
+        $include_nulls: false,
+        value: filter[AgencyFilterKeys.jurisdiction],
+        key: AgencyFilterKeys.jurisdiction,
+        caseSensitive: true,
+      });
+
+      filters.push({
+        OR: builtFilter.map((entry) => ({
+          jurisdictions: {
+            id: entry,
+          },
+        })),
+      });
+    });
+
+    return { AND: filters };
   }
 
   /**
@@ -174,7 +192,7 @@ export class AgencyService {
    *
    * @param agencyId - The ID of the agency which should be removed.
    * @returns A `SuccessDTO` indicating that the delete operation completed successfully.
-   * @throws {BadRequestException} If no property ID is provided.
+   * @throws {BadRequestException} If no agency ID is provided or the agency is associated with users.
    */
   async deleteOne(idDto: IdDTO) {
     if (!idDto || !idDto.id) {
@@ -182,6 +200,18 @@ export class AgencyService {
     }
 
     await this.findOrThrowAgency(idDto.id);
+
+    const associatedUsersCount = await this.prisma.userAccounts.count({
+      where: {
+        agencyId: idDto.id,
+      },
+    });
+
+    if (associatedUsersCount > 0) {
+      throw new BadRequestException(
+        'This agency is currently associated with user(s) and is unable to be deleted.',
+      );
+    }
 
     await this.prisma.agency.delete({
       where: {
